monkeydemo/requirements-capture-screen.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
itsdangerous==0.24
picamera==1.8
pillow
pyscreenshot
